requirements.txt
Django==2.2.1
bcrypt
django-multiselectfield
django-htmlmin
django-otp
django-taggit
git+https://github.com/dokterbob/django-newsletter.git
pycountry
social-auth-app-django
qrcode
arrow
Pillow
Markdown
readtime
psycopg2
gitpython
uwsgi
mysqlclient

# Store
django-paypal
weasyprint
celery
localflavor